change uint to uint64

+++ dialect.go
@@ -62,7 +62,7 @@ type Dialect interface {
 
 	// In case the database returns a row ID, it will resolve the row ID with the primary key of the row
 	// Currently only supports a single primary key
-	ResolveRowID(tableName string, rowID uint) uint
+	ResolveRowID(tableName string, rowID uint64) uint64
 
 	// Separator at the end of a statement
 	ClientStatementSeparator() string
